Add files via upload
The beginning foundations of our new thing editor with a database with a frontend and backend written in JavaScript and HTML and uses Vue which is a competitor to JavaScript react. my first time learning/coding using Vue and a database with a frontend and a backend for a webapp and it is comment heavy for all of us to understand everything in plain English.